fix: Update Cargo.lock with new dependencies
This commit updates the `Cargo.lock` file to reflect changes in
dependencies.

- Added `minimax` plugin. - Added `run-asm` plugin. - Updated `bitflags`
to `2.9.2`. - Updated `cc` to `1.2.33`. - Updated `cfg-if` to `1.0.3`. -
Updated `data-url` to `0.3.2`. - Updated `form_urlencoded` to `1.2.2`. -
Updated `generator` to `0.8.7`. - Updated `hyper` to `1.7.0`. - Updated
`idna` to `1.1.0`. - Updated `proc-macro2` to `1.0.101`. - Updated
`serde_json` to `1.0.143`. - Updated `syn` to `2.0.106`. - Updated
`thiserror` to `2.0.16`. - Updated `tinyvec` to `1.10.0`. - Updated
`url` to `2.5.5`. - Updated `winapi-util` to `0.1.10`.